requirements-dev.txt
3-1==1.0.0
asgiref==3.2.10
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
static3==0.7.0
whitenoise==5.2.0
